--- src/helpers/eventsTimeStamp.js
@@ -1,8 +1,0 @@
-const timeStamp = () => {
-  const date = new Date();
-  const utcDate = date.toLocaleString();
-  // console.log(utcDate);
-  return utcDate;
-};
-
-export default timeStamp;
